Enums are defined as: [enum integer value] [enum value name] [enum description] <: label>
For example:
2 UP This indicates the component is in an UP state: Up

If you specify useNameAsLabel, the above example would result in "UP" being used as the displayed value in the combo box for the enum.

If you don't specify useNameAsLabel, you would see "Up" in the combo box.

If the label isn't specified (and you HAVE NOT specified useNameAsLabel) the description portion of the enum will be used - which you don't generally want.