Initial support for extended reference types that allow for additional information to be stored on object references.
An example definition:

ExtendedReferenceTypeDefinition
name                    SomeRelation
extendedReferenceClass	ObjWithRefs
fieldSeparator :
field Integer count Some kind of counter.
field Integer order The order in which something is done.
description 			The SomeRelation type is used to specify the a count and an ordering
 on some relation between objects. 

It is recommended that a field separator be used - you should ensure that the separator is not used in the name of the object being referenced or in the data fields.

The field attribute has the form:

field <type> <field name> <description>