Manual porting of the following change from 2.0.6 r1393
Breaking down the static initialization of the compact schema into separate
functions didn't have the desired outcome; the Java compiler recombines static
initializations into a single function. The new approach is to generate separate
helper classes schemaAMAPDMSAG and schemaCMAPDMSAG that have a separate static
method that performs initialization of the attribute and class maps. Ideally,
schemas shouldn't grow so large that this becomes an issue, but this should
prevent the static initializers exceeding length limits.