Added support for back reference tracking from complex types.
NOTE: is you're using complex types in combination with backreference tracking, be aware that deletion of an object referred to by a complex type will simpley cause the object reference in the complex type to be set to unresolved. We can't use the regular mechanisms of deleting the attribute (of the complex type) since there may be multiple parts of the type that refer to objects.