Regen for DSD changes.
Addition of the concept of greedy parts for the complex type. This allows for the final requiredPart to grab all remaining text without parsing it or getting confused by it. Primarily added for the example attribute that now features a type a title and everything else.

Added in the ability to retrieve the ConfigLocation for a module in the base generator utility.