Addition of the channel based debug mechanism that allows for fine grain control over access to debug information.
Added support for resolving references in complex types that refer to other complex types that contain references. This necessitated adding a ComplexTypeDefinition handle to internally generated TypeDefinitions so that we could determine if a referred to ComplexType contained references.

Also added further convenience interfaces to generated complex types to determine if an instance of a type has a value for an optional field, the size of mutil-valued fields and direct access to a value at a particular index in a multi-valued field.