Action trigger generation changes.
Action definitions may now refer to attributes will class specific types
i.e. not just DmcObject types. This is allowed because when we generate
the action trigger, we now use specialized functions to generate type
specific DMO interfaces to set/add objects. Previously, we had been
using the code generation mechanisms for full fledged DMOs and these
would wind up sending on the names of objects because they were
interpreted as references.

But, in the case of actions, we want to be able to send full objects,
regardless of where or not they are named objects.