Support for the dm-start starter project.
Cleaning up of unrequired trace messages.

Added the -qwarnings and -qprogress options to dmo and dmw gen
utilities.

Updated the POMAdjuster to write the adjusted pom.xml

Addition of the InitializationInterfaceManager - this will generate an
interface that can be injected into an initialize() method on the base
class of a DSL.  This allows us to generate a config loader for a new
DSL that includes calling initialize() on all loaded objects.
Previously, this had to be implemented by hand - now it becomes a
convention.

This class also generates a ConfigLoader for a DSL, another piece that
used to be created by hand previously. This config loader can now be
used by a generated utility for a DSL created via dm-start.

The DSLArtifactGenerator now generates a base
DSL/tools/DSLutil/DSLUtilMain class that implements a basic skeleton for
the DSL utility - this includes a main() method that is referred to in
the pom.xml as the main class for the shaded jar. In addition, a small
set of test data is generated along with a JUnit test that exercises the
DSL utility. So, right out of the gate, the user has a complete running
utility.

Small tweaks to the meta schema to mark dotName and dmoFromModule as
NONPERSISTENT. Also, when the definedInXXXModule attribute is generated
for a DSL it is also marked as NONPERSISTENT. This allows for
toPersistentOIF() and toJSON() to skip these attributes.

DmwWrapper now has direct support for toJSON().

The DMWGenerator dumpExtendedClass() now generates the standard
initialize() method that will be called by the generated ConfigLoader as
well as the toJSON() method for the DSL's module class so that it's easy
to dump a JSON representation of an entire module; just in case you like
JSON ;-)