Branch 3.1.13
Changes to allow for proper formatting of JSON. Numeric types and
Boolean were being printed with quotes around the values; this indicated
that they were Strings.

DmcAttribute updated to check if an attribute is numeric or boolean and
format accordingly.

DmcAttributeInfo updated to recognize primitive types as numeric or
boolean and to the isNumericOrBoolean() function to check on this
aspect. This approach is used in DmcAttribute.

The metaschema is updated to add numericOrBoolean to TypeDefinitions and
to flag primitive types with this information. This approach is used in
the NewComplexTypeFormatter when formatting complex types.

The NewComplexTypeFormatter uses the type information to determine the
correct form of JSON output to use in the generated toJSON() method of
complex types.